+++ src/js/login.js
@@ -1,57 +1,59 @@
 // login
 
+
+
+
 let searchBox = document.querySelector(".header__box");
 let modalLogin = document.querySelector(".modal__wrapper");
 let userEnter = document.querySelector(".header__user");
 let loginBox = document.querySelector(".header__login-box");
 
-// function submitForm() {
-//   let emailValue = document.getElementById("loginForm").elements.email.value;
-//   let passwordValue =
-//     document.getElementById("loginForm").elements.password.value;
-
-//   // Sign in with email and password
-//   firebase
-//     .auth()
-//     .signInWithEmailAndPassword(emailValue, passwordValue)
-//     .then((userCredential) => {
-//       let user = userCredential.user;
-//       searchBox.classList.add("show");
-//       modalLogin.classList.add("hide");
-//       userEnter.classList.add("hide");
-//       loginBox.classList.add("show-box");
-//       console.log("Login successful:", user);
-//     })
-//     .catch((error) => {
-//       var errorCode = error.code;
-//       var errorMessage = error.message;
-//       console.error("Login error:", errorCode, errorMessage);
-//     });
-// }
 
 function submitForm() {
   let emailValue = document.getElementById("loginForm").elements.email.value;
-  let passwordValue =
-    document.getElementById("loginForm").elements.password.value;
+  let passwordValue = document.getElementById("loginForm").elements.password.value;
 
-  // Увійти за допомогою електронної пошти та паролю
-  firebase
-    .auth()
+  firebase.auth()
     .signInWithEmailAndPassword(emailValue, passwordValue)
     .then((userCredential) => {
       let user = userCredential.user;
 
-      // Зберегти інформацію про користувача в localStorage
-      localStorage.setItem("userId", user.uid);
-      localStorage.setItem("userEmail", user.email);
+      // Зчитати роль користувача з бази даних або Firestore
+      const db = firebase.firestore();
+      const userDocRef = db.collection('users').doc(user.uid);
 
-      // Додати/видалити класи або виконати інші дії, пов'язані з інтерфейсом
-      searchBox.classList.add("show");
-      modalLogin.classList.add("hide");
-      userEnter.classList.add("hide");
-      loginBox.classList.add("show-box");
+      userDocRef.get().then((doc) => {
+        if (doc.exists) {
+          const userRole = doc.data().role;
 
-      console.log("Успішний вхід:", user);
+          // Зберегти роль в localStorage
+          localStorage.setItem('userRole', userRole);
+
+          // Зберегти інші дані користувача в localStorage
+          localStorage.setItem('userId', user.uid);
+          localStorage.setItem('userEmail', user.email);
+          localStorage.setItem('userName', doc.data().firstName); // Припустимо, що в Firestore є поле name
+
+          // Додати/видалити класи або виконати інші дії, пов'язані з інтерфейсом
+          searchBox.classList.add("show");
+          modalLogin.classList.add("hide");
+          userEnter.classList.add("hide");
+          loginBox.classList.add("show-box");
+
+          // Виконайте додаткові дії відповідно до ролі користувача
+          if (userRole === 'admin') {
+            console.log(userRole)
+            // Дії для адміністратора
+          } else {
+            console.log(userRole)
+            // Дії для звичайного користувача
+          }
+
+          console.log("Успішний вхід:", user);
+        } else {
+          console.error('Документ користувача не існує!');
+        }
+      });
     })
     .catch((error) => {
       var errorCode = error.code;
@@ -80,12 +82,20 @@ if (userId && userEmail) {
   );
 }
 
+
 function logOut() {
   localStorage.removeItem("userId");
   localStorage.removeItem("userEmail");
+  localStorage.removeItem("userRole");
+  localStorage.removeItem("userName");
   searchBox.classList.remove("show");
   modalLogin.classList.remove("hide");
   userEnter.classList.remove("hide");
   loginBox.classList.remove("show-box");
   console.log("Користувач вийшов.");
 }
+
+
+
+// test
+
